rename settings to appSettings to avoid collision with ts
