--- spa/jsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "src"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
