+++ spa/src/appSettings.js
