+++ spa/src/components/authentication/util.js
@@ -1,4 +1,4 @@
-import { LS_USER, LS_CSRF_TOKEN } from 'settings';
+import { LS_USER, LS_CSRF_TOKEN } from 'appSettings';
 
 export function handleLoginSuccess(data) {
   localStorage.setItem(LS_USER, JSON.stringify(data.user));
