+++ spa/src/components/donationPage/DonationPage.js
@@ -15,8 +15,9 @@ import {
   GRECAPTCHA_SITE_KEY,
   SALESFORCE_CAMPAIGN_ID_QUERYPARAM,
   FREQUENCY_QUERYPARAM,
-  AMOUNT_QUERYPARAM
-} from 'settings';
+  AMOUNT_QUERYPARAM,
+  CSRF_HEADER
+} from 'appSettings';
 import DonationPageSidebar from 'components/donationPage/DonationPageSidebar';
 import DonationPageFooter from 'components/donationPage/DonationPageFooter';
 import StripePaymentWrapper from 'components/paymentProviders/stripe/StripePaymentWrapper';
@@ -30,8 +31,6 @@ import calculateStripeFee from 'utilities/calculateStripeFee';
 import formatStringAmountForDisplay from 'utilities/formatStringAmountForDisplay';
 import { getFrequencyAdverb } from 'utilities/parseFrequency';
 import { CONTRIBUTION_INTERVALS } from 'constants/contributionIntervals';
-
-import { CSRF_HEADER } from 'settings';
 
 function authorizePayment(paymentData, paymentType, csrftoken) {
   const apiEndpoint =
