+++ spa/src/components/donationPage/LiveDonationPageContainer.js
@@ -13,7 +13,7 @@ import useSubdomain from 'hooks/useSubdomain';
 
 // Analytics
 import { useAnalyticsContext } from 'components/analytics/AnalyticsContext';
-import { HUB_GA_V3_ID } from 'settings';
+import { HUB_GA_V3_ID } from 'appSettings';
 
 // Children
 import SegregatedStyles from 'components/donationPage/SegregatedStyles';
