+++ spa/src/hooks/useSentry.js
@@ -5,7 +5,7 @@ import * as Sentry from '@sentry/react';
 import { Integrations } from '@sentry/tracing';
 
 // Const
-import { SENTRY_DSN_FRONTEND, SENTRY_ENABLE_FRONTEND, ENVIRONMENT } from 'settings';
+import { SENTRY_DSN_FRONTEND, SENTRY_ENABLE_FRONTEND, ENVIRONMENT } from 'appSettings';
 
 /**
  * Typically, just loading sentry at build time is sufficient.
