+++ spa/src/utilities/isAuthenticated.js
@@ -1,4 +1,4 @@
-import { LS_CONTRIBUTOR, LS_USER } from 'settings';
+import { LS_CONTRIBUTOR, LS_USER } from 'appSettings';
 
 function isAuthenticated(for_contributor) {
   if (for_contributor) return localStorage.getItem(LS_CONTRIBUTOR);
